Git-Reference, IO Source, & .CI-CD

+++ API/Database/Models/Mordhau/player.py
@@ -1,6 +1,7 @@
 import sqlalchemy
 
 from sqlalchemy.dialects.postgresql import UUID
+
 from API.Database.Models import ModelBase
 from API.Database.Models import AlcBase
 
@@ -10,7 +11,7 @@ class Player(ModelBase, AlcBase):
 
     player_name = sqlalchemy.Column(sqlalchemy.String, index=True, nullable=False, unique=False)
     playfab_id = sqlalchemy.Column(sqlalchemy.String, unique=True, index=True, nullable=False)
-    discord_id = sqlalchemy.Column(sqlalchemy.Integer, unique=True, index=True, nullable=True)
+    discord_id = sqlalchemy.Column(sqlalchemy.BigInteger, unique=True, index=True, nullable=True)
     team_id = sqlalchemy.Column(
         UUID,
         sqlalchemy.ForeignKey("mfc_teams.id", ondelete="SET NULL"),
